Breakup views into separate files
