+++ admin_ui/views/__init__.py
@@ -1,0 +1,3 @@
+from .change import *
+from .deploy import *
+from .doi import *
